FFDC Improvements to SBE FIFO
In the case where we never see a EOT for a FIFO operation, it
could be useful to see a portion of the data that we received.
This change will add the entire expected response buffer, plus
the first 16 words beyond that, and the last 16 words that we
read out.

Add these registers to most error paths:
 UPFIFO Status
 DNFIFO Status
 Status Register
 CBS Control/Status register
 Selfboot Control/Status register
 Selfboot Message register

Change-Id: Id9e5e7d50cd0db6cad1a9798eca5c42a423c67c9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153594
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>