Purge ATTR_FRU_ID from Hostboot
Some time ago, Hostboot stopped filling in ATTR_FRU_ID during MRW
processing. This meant that the attribute would default to 0 and any
usage of it would always be 0.

In most cases, the code that used that attribute continued working as
expected accidentally. This commit removes the attribute completely in
favor of ATTR_PDR_ENTITY_INSTANCE where applicable.

Change-Id: I63ddaf9e5cfe1348086757672663f770912b2418
JIRA: PFHB-387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153382
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>