Fix issue in target-generation makefile scripts
This commit fixes an issue where makefile targets generated by the
BUILD_GENERATED macro were getting built twice in parallel
builds.

This was due to the fact that the callers of the BUILD_GENERATED
macro were manually adding the generated targets to GEN_PASS_BODY
(i.e. the prerequisite list of the _BUILD/PASSES/GEN/BODY/ACT target)
while the BUILD_GENERATED macro itself was adding the generated
targets to the prerequisite list of GEN_PASS itself. The recursive
'make' invocation in passes.rules.mk was creating another make process
that was unaware of the shared dependency to build the prerequisites
of _BUILD/PASSES/GEN/BODY/ACT while the original 'make' process was
building the other prerequisites of GEN_PASS in parallel, causing the
targets to get built twice.

This commit fixes the issue by altering the BUILD_GENERATED macro so
that it doesn't add the generated build targets to the prerequisite
list of GEN_PASS.

Change-Id: Ifd11c22ca6b586835db3489f1953f85e615ad804
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153600
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>