Add OMI Pipe FIFO for Odyssey
This commit enables the OMI Pipe FIFO for hostboot to
talk to Odyssey. It also does the Pipe Access Control setup,
which is temporary, as SBE will eventually do the setup.
The Pipe FIFO uses the OMI bus and is faster
than the SPPE FIFO. Once the OMI bus is configured,
then an ATTR is set to enable the Pipe FIFO to Odyssey.

Change-Id: I371bbd6aefd7efde72fd62d61e1de8981f3a5432
JIRA: PFHB-262
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151018
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>