Odyssey chipop for ody_thermal_init and polling enablement
This commmit is for adding and enabling Odyssey chipop
for ody_thermal_init and polling. The new chipop sets
the attributes ATTR_ODY_DQS_TRACKING_PERIOD and
ATTR_ODY_SENSOR_POLLING_PERIOD_MS to the desired value
as part of the initialization and enablement. This commit
has changes to send the HWP (Class 0xAC, Command 0x02).
Please note that class 0xAC is for Telemetry and 0x02 is
the control command pertaining to Thermal Sensor and
DQA Polling.
The hostboot code does not support running this HWP
natively for Odyssey, so there is a check to validate
native mode before requesting the HWP and a trace,
if executing in native mode.

Change-Id: I0e1e1e9c7357edfc2cb3674336e4f9736082bf15
JIRA: PFHB-236
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154014
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>