Do not send DDR5 parts list to SBE
This commit has changes to not report to the SBE about any of
the DDR5 Odyssey devices. Typically HB sends a chipop to all
of the processor SBEs with info about the connected OCMBs,
PMICs and GENERIC_I2C_Devices so that SBE can provide
some features such as hardware (scom) access, pmic
telemetry calls and support for hardware dump etc. But
these communications are not required for Odyssey as there
are other mechanisms in place (Example I2CR) to provide
that functionality.

This change will help avoid adding Odyssey specific support
to the P10 SBE code.

Change-Id: Ibe7a58350bc7b58f1f6936c315bed211eb050052
JIRA: PFHB-419
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154246
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>