Add code version info to Odyssey code update error logs
This commit adds Odyssey code version info to code update-related
error logs to help in debugging. Currently the only version
information added is the 64-bit prefix of the SHA512 hash for the
bootloader and runtime code versions. Later we will add the
human-readable code version strings to the logs (see PFHB-491).

Change-Id: I45a24c7dfe46ebb6f05329eba3782f5d050b60af
JIRA: PFHB-261
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155313
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>