Work around register setup issues when enabling refresh
These register bits that were erroneously left untimed in the
design. This means that care needs to be taken to sidestep timing
hazards in the hardware. In general, this means minimizing the number
of bits that change when logic is active. Setting up the configuration
bits before changing the enable bits does this.

In this change, power controls are unconditionally enabled. This means
that the FSM will not try to transition upon force exit self refresh
going low. Instead, it will only transition upon dfi_update/zq
cal. Since those are periodics, they don't immediately engage upon
enablement so they are safe as long as you set the period in a
separate scom.

If the user does not require power controls, they are still enabled,
but they are set with a non-aggressive value so that they will have
negligible effect in performance while still achieving the goal of
fixing these untimed bits.

Change-Id: I27dd80808d217bfbf0c936671216ad478208d34d
EWM: 307274
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154670
Dev-Ready: Louis Stermole <stermole@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Sneha Kadam <sneha.kadam1@ibm.com>
Reviewed-by: Edgar Cordero <ecordero@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154671
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>