IStepDispatcher: Unload modules in reverse load order
This commit changes the IStep Dispatcher so that it unloads modules in
the reverse order of that which it loads them in. This is done so
that, as the initialization/loading of one module can depend on
(i.e. call functions, or reference data, within) a module earlier in
the IStep's dependency list, the destruction/unloading of a module can
depend on a module earlier in the IStep's dependency list.

Change-Id: Ic33f9253095f9d45211b610f333eb6ee0068f4ae
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155314
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>