Process Attr Mismatch Errors from SPPE
This commit introduces changes to process the attribute error
array returned to us from SPPE when we perform attribute sync
(both "push" and "pull"). If any errors are returned, an informational
error log is created and committed, containing all of the attribute
errors communicated by SPPE in its FFDC fields. It was agreed that
this log be informational and not fail the boot in any scenarios.

Change-Id: Id6d208780bffac582466cac1eb477e29aa5828f1
JIRA: PFHB-439
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155037
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>