SBE FIFO in HBRT
This commit commonizes the SBE FIFO access logic between HBRT and
IPL-time HB. This logic is used to perform SPPE chip-ops during channel
failures (when no other way of communication with SPPE is available).
The core of the logic is the same, so the actual FIFO code hasn't
changed.

In this commit:
- Various makefile changes to include the FIFO logic into HBRT.
- An increase in HBRT PNOR partition to accommodate the new code size.
- A couple of fixes for testcases that exercise FIFO logic in HBRT.

Change-Id: Id20c6e367a4805349e37e042a7f541dc608378d4
JIRA: PFHB-400
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155766
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>