Fix error leak in HDAT when failing to get NVRAM section
While HDAT is attempting to run hdatFillDataPtrs() there is an attempt
to get section info for the NVRAM section. In an error case, the log
which is returned from that attempt is not properly handled leading to a
leak when the handle is reused later in the function and overwritten.

Change-Id: Ie99a3549d671bf590f1167c9e04a936c8f4dbc23
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155718
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>