Purge ATTR_FRU_ID from HDAT
Some time ago, Hostboot stopped filling in ATTR_FRU_ID during MRW
processing. This meant that the attribute would default to 0 and any
usage of it would always be 0.

In most cases, the code that used that attribute continued working as
expected accidentally. This commit removes the attribute completely in
favor of ATTR_PDR_ENTITY_INSTANCE where applicable.

Change-Id: Ic71978c2bb3766fe508402643166cc7246f93520
STGD:561609
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155987
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>