Introduce deactivated discontiguous memory allocator
This commit introduces the code for a discontiguous memory allocator
to implement malloc() and free(). The allocator is not active yet, and
all code still uses the contiguous allocator. (See below.)

A discontiguous allocator is useful because for most cases, Hostboot
allocates memory and doesn't expect the physical pages that back the
memory range to be contiguous. Allowing the physical pages to be
discontiguous helps us deal with memory fragmentation by allowing large
allocations to be split up wherever they can fit in the physical
address space.

There are certain places where contiguous physical memory is required,
such as by the kernel, or in communication with the SBE, which reads
physical memory directly without going through page translation. These
places call the contiguous-backed allocator explicitly.

This allocator allocates a block of virtual memory space, and uses a
"bump" allocator strategy (i.e. new allocations always come from the
front of the heap, and allocations are never reused).

The allocator is not activated yet because we are still experimenting
with the best heap sizes and allocation strategies.

JIRA: PFHB-521
Change-Id: I7e8de943e0dc348845a2fa3d9beb860bed92a99e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155995
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>