Do not set gard/deconfig bit for Clocks
BMC's NAG functionality exposed an issue with error logs associated with
clocks. When a guard even is requested for a clock, we will set the
"guarded" bit in the error log, but we won't actually guard the part,
since we don't support clock guards. This commit removes setting of
deconfigured/guarded error log bits for clock targets so that the NAG
tools that looks at them doesn't get confused.

Change-Id: Ida000f28ed9fb4612414431e7688728661f56679
STGD: 560540
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156441
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>