Increase reset timeout for NV Controller
The latest level of NV firmware takes longer to come out of
reset than we allow.  Currently we take the value from a SPD
byte (5 on the parts we have) and multiply it by 2 to be safe.
The new firwmare takes 12 seconds.  This change adds an
additional 30 seconds to the timeout to cover any future
increases.

Change-Id: I92e5de85a931b8c6db420289b948d55774328baf
STGD: 551313
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156099
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>