Odyssey code update: Fix sides out-of-sync problem
BACKGROUND
==========

The Odyssey SBE has two EEPROMs or "sides" that each stores a copy of
the bootloader and runtime code images (so, four images total).

The Odyssey SBE does not offer any way to obtain the hashes of the
code images on the side that is not currently running. Hostboot can
only tell what code level is on the running side.

With this in mind, a high-level slice of the Odyssey code update
process (before this commit) works as follows:

1. Boot an Odyssey up using side 0 code levels.
2. Check whether the runtime or bootloader images are out of sync with
   what Hostboot has.
3. If either are out of sync, write Hostboot's versions to the side 1
   copies of the image(s) that are out of sync. Note that we do not
   update an image on side 1 if that image on side 0 is in sync with
   Hostboot.
4. Restart the Odyssey, this time booting using the just-updated
   code levels on side 1.
5. The code levels on side 1 should now match what Hostboot has. If
   they do not match, the code update failed and the OCMB is assumed
   to be faulty.

PROBLEM
=======

There is an issue with this algorithm, which is that if side 1 has an
out-of-date bootloader and side 0 has an up-to-date bootloader,
Hostboot will not update the bootloader on side 1.

Then, when Hostboot boots the Odyssey from side 1, it will see that
the code levels are still out of sync after a code update, and it will
assume that the OCMB is bad.

POTENTIAL SOLUTIONS
===================

There are at least two potential solutions:

1. Update both images when either image is out of sync.
   - Problem: This may cause us to update an image that doesn't need
     to be updated, which adds time to the boot.
2. Call syncCodeLevels to write the active side's code levels to other
   side, just before updating the other side. This ensures that
   whatever determinations we make about side 0 also apply to side 1.

CHOSEN SOLUTION
===============

This commit implements option 2, because it handles all cases while
optimizing for the good case. syncCodeLevels is very fast over I2C and
OMI. It shouldn't ever overwrite a good image with a bad image,
because the image being written will have booted us up to a point
where the FSM thinks it is fastest to perform the code update.

We may revisit this if the SBE ever provides a way to view off-side
code levels.

Change-Id: I2f7d314fe53dd3f79f142d5d4cc7318ee92a9b80
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156168
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>