Hack to workaround SBE sending RC=0 FFDC packages
Presently, SBE is sending back RC=0 FFDC packages when the SBE does not
fail. Sometimes these packages can have data in them but there is no way
to tell if it's a PLAT or HWP FFDC package. For now, stuff the data in
the PLAT log and do not go down the HWP processing path.

If the HWP processing path is taken, RC=0 will generate an invalid
packet error which confuses people.

Change-Id: I9ab84f1d2855de829bc5f753fe18af512bd864d2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156238
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>