Deliver dce tools and metadata to the BMC filesystem
This commit delivers dce-bmc-invoke.sh and preplib.py to the BMC
filesystem, and creates hb_tooldata.tar.gz which will become a LID on
BMC systems. This LID is intended to contain whatever extra
image-dependent debug data our tools need. As of this commit, it will
contain mangled symbol information for Hostboot images.

The intended use for these files is to allow DCE binaries to be linked
on the system where they will run, which will obviate the need to
recompile DCE binaries for each driver/system in most cases.

Change-Id: If3bc3beb0f70dc0f21119f9b1460838e5049344b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156243
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>