Use decimal numbers for BPM version header
The logic inside the BPM uses a hex representation of a decimal
value for the versioning.  Up to this point we've used a decimal
value stored as a binary hex.  A new level of BPM code is the
first to go to 2 decimal digits which exposed our mismatched
assumption.

Specifically, v1.11 of the BPM code is represented as 0x0111 in
the BPM itself, but our packaging script was using 0x010B.

This change will reconcile our logic to match the empirical
results from the BPM.

Change-Id: I97b54a995c8152ea0341dfe11ec6e47fefbc5250
STGD: 551313
Jira: PFHB-510
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156569
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>