Fix lost error paths in proc_build_smp istep
There are a number of error paths in the proc_build_smp istep
code that end up being ignored and allowing the IPL to continue
when it shouldn't.  Besides losing valid errors it can also
lead to confusing errors inside subsequent isteps because of
the code that is skipped.

Change-Id: Ia20212cd63e64a9b84881eba4bae638d6ff2ffd1
STGD: 567504
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156640
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>