TPM Extend Hash Resiliency
Allow for recovery from possible aborted sequences which may occur
during the transition from PHYP to Hostboot during an MPIPL.

The TPM STS register may contain an 0xFF value which is NOT a valid
condition, which may falsely represent conditions which are
undesirable.  Catch the occurrence of 0xFF during the tpmReadSTSReg
and poll for a valid signature combination.

Change-Id: Ica20d79b41d7de2cba855ada6a77cc090890df6b
STGD: 489295
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156646
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>