Improved FFDC for NVDIMM
This commit contains several improvements to the FFDC that
is collected for NVDIMM errors and significant events.

Hostboot creates an informational log at the start and end of
the NVDIMM code update flow.  The SRC includes the HUID so we
know which position is being updated but we don't know which
specific instance.  Adding a FRU callout to the informational
log should cause the log to include the serial and part number
so that a specific instance can be identified.

A change was made in a previous NV firmware release that prevents
vendor logs from being collected before the part is unlocked.
This is a problem because the checks for CSAVE errors happen
before the part is unlocked.  This change moves the unlock earlier
so that all of the FFDC can be collected in the fail case.

The vendor log from the NVDIMM can be quite large but we have a
limit on the size of error logs that are sent from HBRT to the
FSP.  This change breaks up the log into smaller pieces so that
some data is more likely to fit into the original error.  It will
also put the entire log into multiple additional informational pels.

Some recently discovered changes to the vendor log format have
broken the logic we use to save this data off.  Changes have
been made to accommodate the various scenarios that the log
could appear in.  A change was also made to replace all NULL
characters with a printable string "~NULL~" to allow the data
to be saved and displayed properly.

A new command was added to the HBRT runtime interface that
can be used to save the vendor logs out to a PEL at runtime.

Change-Id: I5d86a940d342e9ee9beda989c2659545193e27e7
STGD: 570867
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/157648
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>