Migrate guard records to handle new DIMM numbering
Changes required to support Odyssey/DDR5 in FW1060 that resulted
in some of the numbering schemes (specifically PHYS_PATH) having
to be modified to support double the number of memory ports behind
the OCMB. What that means is that if a system were to have a guard
record present for a (logical) DIMM then that system was updated
to the level of code with this change, the meaning of that guard
record would change.  For example, a guard of DIMM4=C44 would
mutate to be a guard of (new)DIMM4=C45. The same thing would
happen going the other direction.

To solve this problem we will increment the version number of the
guard record format in FW1060.  The version will be used as the key
to know if we need to transform data from an existing guard record
into the numbering scheme that we now have.  Similarly we will need
to reverse this transformation to support code downgrades
(1060->1050).

Change-Id: I844ede0d30d315ed35a3e70cc2dcb3e972b0c7f7
JIRA: PFHB-538
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/157722
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>