Disable BPM register polling during update flow
The recently introduced v4.4 NV code has a bug that can cause BPM
segment data to be corrupted during BPM firmware update. The bug
occurs when the newly-introduced NVDIMM polling of BPM registers
for register mirroring interferes with reading the old segment
data from the BPM at the beginning of a BPM firmware update,
causing the wrong values to be read and subsequently written back
to the BPM after the BPM firmware update completes.

Change-Id: I3c2fa88dc2d3820a33216ca545f90510f57366c3
STGD: 573387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/157649
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>