Update boot proc SBE if it's downlevel during step 7.5
In some field replace scenarios it is possble for the boot proc FRU to
be replaced with hardware that is downlevel compared to the system. This
could lead to errors as the system attempts to do an SBE update in 10.2.

To properly address this situation, code has been added in 7.5 to update
downlevel boot proc SBEs to avoid those errors.

Change-Id: I32f58e5c39c62b7aa1d261d2cf435dc0748b5f90
JIRA: PFHB-564
STGD: 566189
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/157677
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>