Fix makefile dependency to re-enable building in subdirectories
Previously, the toplevel makefile would add targets to the
FINAL_PASS variable; everything in FINAL_PASS depends on the rest
of the code; and the "all" target would depend on FINAL_PASS.
This ensured that everything in FINAL_PASS would get built after
the rest of the code/images/etc.

When building from a subdirectory, nothing is added to FINAL_PASS,
so a `make` invocation would do nothing because the "all" target
had no prerequisites.

This commit fixes the issue by making the "all" target have a
dependency on the rest of the code as well as the things in
FINAL_PASS. This way, even when FINAL_PASS is empty, a `make`
invocation will actually build the code.

Change-Id: If72951320c28688ac99307f1d7e94187c6373c15
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/157902
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>