Support TPM-less Boot
This commit fixes an issue seen on a system with TPM physically not
installed.
Without a physical TPM, SBE measurement code will attempt to read the
TPM version and will fail, setting the TDP (TPM Deconfig Protect) bit
but also not releasing the lock on the SPI TPM engine. Hostboot will not
perform any TPM ops because the TDP bit is set, but it will attempt to
release the SPI lock, and that operation will fail.

This commit avoids attempting to unlock the SPI TPM engine if there is
no functional primary TPM present in the system. Note however, that SBE
SMP build flow still requires us to enter and exit TPM extend mode, so
those steps need to be run.

Change-Id: Id1d993d6c8e48ef7ec1f33bdbd6847c3039c9798
JIRA: PFHB-597
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158061
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>