Skip ocmb i2c security logic on odyssey
The p10_disable_ocmb_i2c function is currently used to disable i2c
access to the device address used by Explorer and Odyssey. This was
required for Explorer but Odyssey has security function built in,
and this interface is needed for some failure recovery cases. This
commit skips the p10_disable_ocmb_i2c function when Odyssey OCMBs
are detected. This assumes there will not be mixed OCMB types on
the system, which will be verified by a different commit.

Change-Id: I5c4e80bde6a2dcba62f55493d2e67627c777c6ba
JIRA:PFHB-509
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158167
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>