Check inputs for hb-pauseIstepsAt
The standalone simics tool hb-pauseIstepsAt takes two args, a major and
minor istep. There was an issue where if the user didn't supply either
arg or just major istep the code would try to write blank strings into
the attribute responsible for the pause. This caused a python error.

To prevent this, this commit updates the code to check for blank args
and handles them.

Change-Id: I59e35b81fb84a35f81556cd5ba87643f2deaca48
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158177
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Ilya Smirnov <ismirno@us.ibm.com>