Fix DCE handling in xlate_stack
Relocates where the checks are for the DCE handling so that it can
capture scenarios where the addresses are smaller than the "give-up"
threshold for the normal path. This avoids some situations in which the
normal path will attempt to decode DCE addresses and give non-sense back
to the user.

Change-Id: I056d25f1b52517e79e91f1969230c163cb2ce54e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158229
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>