Enable PLDM query of BMC Reboot Attempts Left
Enable sendGracefulRestartRequest to query the BMC boot
attempts left in order to properly handle selectively
adding a BMC boot attempt to allow reconfig loops to properly
decrement the final BMC boot count during recovery flows.

If the BMC query of the AttemptsLeft is UNSUPPORTED, default
action is to skip altering the boot count to allow a normal
decrement to eventually reach zero.

Any attempts by Hostboot to perform a graceful reboot will be
honored by the BMC until such time that a TI/Crash/Checkstop
will then be evaluated by the BMC and if the AttemptsLeft is
zero, then the auto reboot will be exhausted and no more
boot retries will be performed.

Change-Id: I3a8f749a64649a27d33305fc7e30dbb6ba929bdb
STGD: 567510
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158310
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>