Reconfig Loop on Backing Cache Update
In a scenario where there is not enough backing cache, Hostboot
terminates with the SHUTDOWN_REQUESTED_BY_FSP code (in FSP path).
However, FSP didn't actually request the shutdown, so when they get
that RC in this situation, they just terminate instead of rebooting
Hostboot. This commit makes a change that asks FSP to do a reconfig
loop in the scenario where there is not enough backing cache.

Change-Id: Ia790b2213c365232559b95322525c4f1e6d2371b
STGD: 525387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170594
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>