Add workaround for ODPCTRL scom bug
There is a hardware bug in Odyssey that requires a firmware workaround.
Scoms to the ODPCTRL register can be corrupted.  This doesn't currently
affect any mainline code but it does impact RAS testing.

This commit adds a workaround when Odyssey targets attempt to read
0801300a/0801340a by doing an indirect scom to the first functional
memport target, discarding that data, and proceeding with the read.

Change-Id: I8cf68ed2b7c8d52cce5647d682769ea0baf72fae
JIRA:PFHB-473
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158453
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>