Fix cross-module TLS references when there are no intra-module references
There was previously an issue with Hostboot's custom linker, where it
would not consider a symbol to be a TLS symbol unless there was an
intra-module reference to it. For example, if a module defined a TLS
variable but never read from or wrote to this variable itself, the
custom linker would consider it to be a normal variable. This would
cause it to emit incorrect relocations, which manifested as crashes
during Hostboot IPLs.

This commit fixes the issue by adding the symbol's name to the TLS
table if it is marked as TLS in the module's symbol list (as opposed
to depending on information from its relocation list).

It also fixes a small bug where the runtime TLS relocation code crashes
if a TLS module ID of 0 appears in a relocation. The bug is due to the
fact that the code uses a 32-bit comparison between 0 and "TLS\0\0\0\0\0"
(which is a TLS module relocation with a module ID of 0 in the lower
32 bits and the "TLS\0" tag in the upper 32 bits). Since the lower
32-bits are 0, the comparison sees equality, which causes the code to
process the relocation incorrectly.

This bug was fixed by starting TLS module IDs at 1.

Change-Id: If3ecfefecab24ad7d46f6eeadc79159cc9cb5288
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158516
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>